move util function and test
