+++ src/components/HistoryTable/hooks/useTestResults.test.tsx
@@ -62,6 +62,7 @@ describe("useMergedHookRender - sanity check", () => {
       pageCount: 0,
       columnLimit: 7,
       commitCount: 10,
+      onChangeTableWidth: expect.any(Function),
     });
   });
 });
